+++ server.h
@@ -4,11 +4,9 @@
 //  单个服务节点的数据结构
 struct Server
 {
-    int index;
+    size_t index;
     int capacity;
     int gpu_used;
-    std::vector<int> assigned_tasks;
-    Server(int i = 0, int c = 0, int w = 0) : index(i), capacity(c), gpu_used(w) {}
+    std::vector<size_t> assigned_tasks;
+    Server(size_t i = 0, int c = 0, int w = 0) : index(i), capacity(c), gpu_used(w) {}
 };
-
-
