+++ task.h
@@ -3,9 +3,9 @@
 // 任务信息的数据结构
 struct Task
 {
-    int index;
-    int start_node;
-    int current_node;
+    size_t index;
+    size_t start_node;
+    size_t current_node;
     int demand;
-    Task(int i = 0, int s = 0, int g = 0) : index(i), start_node(s), current_node(s), demand(g) {}
+    Task(size_t i = 0, size_t s = 0, int g = 0) : index(i), start_node(s), current_node(s), demand(g) {}
 };
